Added scenegraph. Cleaned up app, console and log classes a little. Added sampletime_t for storing the time that the application has been running, frame time, etc.

+++ library/src/breathe/gui/cWidget.cpp
@@ -68,8 +68,8 @@ namespace breathe
 
 		cWidget::~cWidget()
 		{
-			unsigned int n = child.size();
-			for (unsigned int i = 0; i < n; i++)
+			size_t n = child.size();
+			for (size_t i = 0; i < n; i++)
 				SAFE_DELETE(child[i]);
 		}
 			
@@ -85,8 +85,8 @@ namespace breathe
 		{
 			if (_id == id) return this;
 
-			unsigned int n = child.size();
-			for (unsigned int i = 0; i < n; i++)
+			size_t n = child.size();
+			for (size_t i = 0; i < n; i++)
 				child[i]->FindChild(_id);
 
 			return NULL;
