+++ library/src/breathe/render/cFont.cpp
@@ -315,7 +315,7 @@ namespace breathe
 				//  If you decide to use it make sure to also uncomment the glBitmap command
 				//  in make_dlist().
 				//	glRasterPos2f(0,0);
-				glCallLists(lines[i].length(), GL_UNSIGNED_BYTE, lines[i].c_str());
+				glCallLists(static_cast<unsigned int>(lines[i].length()), GL_UNSIGNED_BYTE, lines[i].c_str());
 				//	float rpos[4];
 				//	glGetFloatv(GL_CURRENT_RASTER_POSITION ,rpos);
 				//	float len=x-rpos[0];
